+++ CHANGELOG.markdown
@@ -5,6 +5,7 @@
   - `Data.Functor.Alt.(<!>)` -> `Data.Functor.Alt.galt`
   - `Data.Functor.Apply.{liftF2,liftF3}` -> `Data.Functor.Apply.{gliftF2,gliftF3}`
   - `Data.Functor.Bind.(>>-)` -> `Data.Functor.Bind.gbind`
+  - `Data.Functor.Contravariant.Conclude.{conclude,concluded}` -> `Data.Functor.Contravariant.Conclude.{gconclude,gconcluded}`
 
 5.3.7 [2022.01.09]
 ------------------